Новая пошаговая игра #187 - Рефакторинг 1
