+++ frontend/src/app/AppView.jsx
@@ -1,37 +1,14 @@
+import {GameLobby} from '@pages/game-lobby/GameLobby';
+import {Game} from '@pages/game/Game';
 import {NotFound} from '@pages/not-found/NotFound';
-import {GameType} from '@shared/lib/game/GameType';
-import {SocketProvider} from '@shared/lib/socket/SocketProvider';
-import {Route, Routes, useParams} from 'react-router-dom';
-
-const GamePage = () => {
-  const {gameType, gameId} = useParams();
-  const Game = GameType[gameType];
-  if (!Game) return <NotFound/>;
-  const Provider = Game.provider;
-  const Component = Game.component;
-
-  return (
-    <SocketProvider>
-      <Provider gameId={gameId}>
-        <Component/>
-      </Provider>
-    </SocketProvider>
-  );
-};
-
-const GameLobbyPage = () => {
-  const {gameType} = useParams();
-  const Game = GameType[gameType];
-  if (!Game) return <NotFound/>;
-  return Game.lobby();
-};
+import {Route, Routes} from 'react-router-dom';
 
 export const AppView = () => (
   <Routes>
     <Route path="/login/success" element={<>Успех</>}/>
     <Route path="/login/failure" element={<>Провал</>}/>
-    <Route path="/game/:gameType" element={<GameLobbyPage/>}/>
-    <Route path="/game/:gameType/:gameId" element={<GamePage/>}/>
+      <Route path="/game/:gameType" element={<GameLobby/>}/>
+      <Route path="/game/:gameType/:gameId" element={<Game/>}/>
     <Route path="*" element={<NotFound/>}/>
   </Routes>
 );
