+++ frontend/src/entities/game/index.js
@@ -1,2 +1,3 @@
 export {Background} from './ui/Background';
 export {WordTile} from './ui/WordTile';
+export {GameLobbyForm} from './ui/GameLobbyForm';
